Rename `buffer_sweep` module to `buffer_sweep_encoder`.
