+++ crates/wordchipper/src/encoders/token_span_encoder/span_encoders/buffer_sweep_encoder.rs
