+++ crates/wordchipper/src/encoders/token_span_encoder/span_encoders/mod.rs
@@ -1,6 +1,6 @@
 //! # [`SpanEncoder`] Implementations
 
-mod buffer_sweep;
+mod buffer_sweep_encoder;
 mod merge_heap_encoder;
 mod priority_merge_encoder;
 mod span_encoder;
@@ -8,7 +8,7 @@ mod span_encoder_selector;
 mod tail_sweep_encoder;
 
 #[doc(inline)]
-pub use buffer_sweep::*;
+pub use buffer_sweep_encoder::*;
 #[doc(inline)]
 pub use merge_heap_encoder::*;
 #[doc(inline)]
